backend/requirements.txt
langgraph
langchain
streamlit
pandas
python-dotenv
langchain-groq
langchain-google-genai
requests
pydantic
fastapi
uvicorn[standard]
python-multipart
supabase
